Minor improvements in help and fixes regarding working with files other than .dproj
CHG: Modify the way the helps for commands are written to be better. Add a space line between each option.
ADD: Support for "\n" in the help description of parameters or options so it is better looking when reading instead of long line with idea cut anywhere.
FIX:The overall application was removing empty line when the option --DeleteLine was not used. This was because "FdeleteLine" was set to empty so all empty lines matches that! Visibly, this application was used so far only with .dproj file or with file with no empty line!
FIX:Visibly the same comment as above regarding options --ChangeFrom, --RegExReplaceFrom and --InsertAfter.
FIX:We needed to unquote the parameter passed to option "GroupOptions".
ADD:Empty line in "GroupOptions" provided file will simply be ignored instead of generating errors.
ADD:Lines beginning with "//" in provided file for option "GroupOptions" will not cause errors anymore and will simply be ignored.
CHG:The keywords for the options are now coming from ONE list and are define at ONE place. This way, no error of spelling. It adds little complexity for writing a little and allow other kind of errors to happen. But at least, it won't be spelling error and we won't keep rewriting the same magic word everywhere.
RMV:Remove the option to show command line, to show help and register log.