Improve presentation, error reporting and backup file.
ADD: In case of error, we now better reporting where the error has been found. When we were working with simply options and parameters on one single command line, it was not complicated to report where the error was found. But now, with the fact we may now have to process many options in one shot, it was complication to figure where the error was actually located.
ADD: During the process, we now have a better reporting of the actions that currently take place. Things are aligned.
ADD: It will now *always* save a backup file. We will simply add the extension ".bak" at the end of original filename. If it always exists, it will be ".bak(2)", ".bak(3), etc. until we find one that does not already exist.