Improvement regarding the file list in parameter
ADD: The empty lines and lines beginning with "//" in a specified file list in parameter will now be ignored.
ADD: All the files to process will be idendified and located first prior to process them. This will, if we have two file list specified for example, we will not see a sequence of x of 4 and then x of 6. We will see simply one sequence of x of 10.
ADD: We now better report the errors related with specified files. If no files fit a specified filter, it will now be reported as an error.
ADD: If a file list for parameter does not exist, it will now be reported as an error.
ADD: The error reported regarding file lists are now better reported when things are nested together.