Circular reference with --GroupOptions detected
RMV: Remove some unused code that were probably used years ago related with "DrpojEdit" but not related with this project.
ADD: We now better detect and report error related with situation where we would have circular reference with --GroupOptions. It means when it includes a file that will include something that is already included and still currently under process.